requirements.txt
azure-storage-blob
azure-identity
azure-mgmt-resource
azure-mgmt-storage
pandas
python-dotenv